+++ MyGame/BossEnemyFollow.cpp
@@ -96,10 +96,18 @@ void BossEnemyFollow::Update(Enemy* enemy)
 
 
 	/*2ˆø”F‘Ì—ÍÝ’è(Ý’è’lˆÈ‰º‚È‚Á‚½‚ç‚Rˆø”‚ÌUŒ‚‚Ö)*/
-	AttackSelect(enemy, Percent::GetParcent(enemy->GetMaxHP(), enemy->GetHP()) <= 30.0f, enemy->HALF_1);
-	AttackSelect(enemy, Percent::GetParcent(enemy->GetMaxHP(), enemy->GetHP()) <= 90.0f, enemy->CIRCLE_1);
-	AttackSelect(enemy, Percent::GetParcent(enemy->GetMaxHP(), enemy->GetHP()) <= 70.0f, enemy->Beam);
-	AttackSelect(enemy, Percent::GetParcent(enemy->GetMaxHP(), enemy->GetHP()) <= 60.0f, enemy->KNOCK);
+	AttackSelect(
+		enemy, Percent::GetParcent(static_cast<float>(enemy->GetMaxHP()), static_cast<float>(enemy->GetHP())) <= 30.0f,
+		enemy->HALF_1);
+	AttackSelect(
+		enemy, Percent::GetParcent(static_cast<float>(enemy->GetMaxHP()), static_cast<float>(enemy->GetHP())) <= 90.0f,
+		enemy->CIRCLE_1);
+	AttackSelect(
+		enemy, Percent::GetParcent(static_cast<float>(enemy->GetMaxHP()), static_cast<float>(enemy->GetHP())) <= 70.0f,
+		enemy->Beam);
+	AttackSelect(
+		enemy, Percent::GetParcent(static_cast<float>(enemy->GetMaxHP()), static_cast<float>(enemy->GetHP())) <= 60.0f,
+		enemy->KNOCK);
 
 	//Ž€–S
 	if (enemy->GetHP() <= 0.f)
